requirements.txt
numpy==1.19.3
scipy
Pillow
cython
matplotlib
scikit-image
tensorflow==1.13.1
keras==2.2.4
opencv-python
h5py==2.10.0
imgaug
IPython[all]